Fixed gcc compilation warnings
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@3853 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ src/media_tools/mpd.c
@@ -522,7 +522,7 @@ static GF_Err gf_mpd_parse_adaptation_set(GF_List *container, GF_XMLNode *root)
 	set->base_URLs = gf_list_new();
 	set->representations = gf_list_new();
 	/*assign default ID and group*/
-	set->id = (u32) set;
+	set->id = (u64) set;
 	set->group = set->id;
 
 	e = gf_list_add(container, set);
